requirements.txt
pandas==1.4.1
scikit-bio==0.5.6